\name{01_Miranda}
\alias{example01}
\docType{data}

\title{Example 1. Frequency of seeing}
\description{A flash of light of variable intensity was presented repeatedly at a fixed location in the visual field of a subject who reported whether the flash was visible.}

\usage{data( "01_Miranda" )}

\format{Data frame with three columns for 10 stimulus levels:
\code{x}: Stimulus level.
\code{r}: Number of successes.
\code{m}: Number of trials.} 

\references{Miranda, M. A. \& Henson, D. B. ``Perimetric sensitivity and response variability in glaucoma with single-stimulus automated perimetry and multiple-stimulus perimetry with verbal feedback'', Acta Ophthalmologica, 86, 202-206, 2008.}

\keyword{datasets}