\name{02_Levi}
\alias{example02}
\docType{data}

\title{Example 2. Visual detection of path deviation}
\description{The subject was presented with the image of a dot moving rightwards on a linear path until it reached the midline of the display, when it changed direction either upwards or downwards. The subject had to indicate the direction.}

\usage{data( "02_Levi" )}

\format{Data frame with three columns for 7 stimulus levels:
\code{x}: Stimulus level.
\code{r}: Number of successes.
\code{m}: Number of trials.}

\references{Levi, D. M. \& Tripathy, S. P. ``Is the ability to identify deviations in multiple trajectories compromised by amblyopia?'', Journal of Vision, 6(12), 1367-1379, 2006.}

\keyword{datasets}