\name{03_Carcagno}
\alias{example03}
\docType{data}

\title{Example 3. Discrimination of pitch}
\description{The subject had to identify the interval containing a tone whose fundamental frequency was different from that in the other two intervals.}

\usage{data( "03_Carcagno" )}

\format{Data frame with three columns for 8 stimulus levels:
\code{x}: Stimulus level.
\code{r}: Number of successes.
\code{m}: Number of trials.}

\references{Unpublished data from S. Carcagno, Lancaster University, July 2008}

\keyword{datasets}