\name{04_Xie}
\alias{example04}
\docType{data}

\title{Example 4. Discrimination of 'porthole' views of natural scenes}
\description{The subject was presented with a display split into two parts, one containing a pair of patches from the same image, the other a pair from different images, and the subject had to judge which pair came from the same image.}

\usage{data( "04_Xie" )}

\format{Data frame with three columns for 10 stimulus levels:
\code{x}: Stimulus level.
\code{r}: Number of successes.
\code{m}: Number of trials.}

\references{Xie, Y. \& Griffin, L. D. ``A 'portholes' experiment for probing perception of small patches of natural images'', Perception, 36, 315, 2007.}

\keyword{datasets}