\name{05_Schofield}
\alias{example05}
\docType{data}

\title{Example 5. Induction of a visual motion aftereffect}
\description{The subject was presented with a moving adaptation stimulus, followed by a test stimulus.}

\usage{data( "05_Schofield" )}

\format{Data frame with three columns for 7 stimulus levels:
\code{x}: Stimulus level.
\code{r}: Number of successes.
\code{m}: Number of trials.}

\references{Schofield, A. J., Ledgeway, T., \& Hutchinson, C. V. ``Asymmetric transfer of the dynamic motion aftereffect between first- and second-order cues and among different second-order cues'', Journal of Vision, 7(8), 1-12, 2007.}

\keyword{datasets}