\name{06_Nascimento}
\alias{example06}
\docType{data}

\title{Example 6. Discrimination of image approximations}
\description{The subject was shown an image of a natural scene and an approximation of this image based on principal component analysis. The task was to distinguish between the images.}

\usage{data( "06_Nascimento" )}

\format{Data frame with three columns for 8 stimulus levels:
\code{x}: Stimulus level.
\code{r}: Number of successes.
\code{m}: Number of trials.} 

\references{Nascimento, S.M.C., Foster, D.H., \& Amano, K. ``Psychophysical estimates of the number of spectral-reflectance basis functions needed to reproduce natural scenes'', Journal of the Optical Society of America A-Optics Image Science and Vision, 22 (6), 1017-1022, 2005.}

\keyword{datasets}