\name{07_Baker}
\alias{example07}
\docType{data}

\title{Example 7. Auditory detection of a gap in noise}
\description{A 300-ms noise burst containing a gap of 2--8 ms duration or no gap was presented to one ear of a subject.}

\usage{data( "07_Baker" )}

\format{Data frame with three columns for 8 stimulus levels:
\code{x}: Stimulus level.
\code{r}: Number of successes.
\code{m}: Number of trials.}

\references{Baker, R. J., Jayewardene, D., Sayle, C., \& Saeed, S. ``Failure to find asymmetry in auditory gap detection'', Laterality: Asymmetries of Body, Brain and Cognition, 13, 1-21, 2008.}

\keyword{datasets}