\name{checkinput}
\alias{checkinput}

\title{Check input parameters of modelfree functions}

\description{THIS IS AN INTERNAL FUNCTION ONLY. Pool of routines that check robustness of input of parameters passed to the other functions of modelfree package.}

\usage{checkinput(type,x)}

\arguments{
\item{type}{ Type of checking }
\item{x}{ Input data to be checked }
}

\value{None. Only stops execution if data is not consistent.}

\keyword{utilities}
