% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comploglog_link.R
\name{comploglog_link}
\alias{comploglog_link}
\title{Complementary log-log link with guessing and lapsing rates
Creates a complementary log-log link function; the guessing rate and lapsing rate are fixed, hence link is
a function of only one variable.}
\usage{
comploglog_link(guessing = 0, lapsing = 0)
}
\arguments{
\item{guessing}{(optional) guessing rate; default is 0}

\item{lapsing}{(optional) lapsing rate; default is 0}
}
\value{
link complementary log-log link for use in all GLM functions
}
\description{
Complementary log-log link with guessing and lapsing rates
Creates a complementary log-log link function; the guessing rate and lapsing rate are fixed, hence link is
a function of only one variable.
}
