% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_link.R
\name{logit_link}
\alias{logit_link}
\title{Logit link with guessing and lapsing rates
Creates a complementary logit link function; the guessing rate and lapsing rate are fixed, hence link is
a function of only one variable.}
\usage{
logit_link(guessing = 0, lapsing = 0)
}
\arguments{
\item{guessing}{(optional) guessing rate; default is 0}

\item{lapsing}{(optional) lapsing rate; default is 0}
}
\value{
link logit link for use in all GLM functions
}
\description{
Logit link with guessing and lapsing rates
Creates a complementary logit link function; the guessing rate and lapsing rate are fixed, hence link is
a function of only one variable.
}
