% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglog_link.R
\name{loglog_link}
\alias{loglog_link}
\title{Log-log link with guessing and lapsing rates}
\usage{
loglog_link(guessing = 0, lapsing = 0)
}
\arguments{
\item{guessing}{(optional) guessing rate; default is 0}

\item{lapsing}{(optional) lapsing rate; default is 0}
}
\value{
link log-log link for use in all GLM functions
}
\description{
Creates a log-log link function; the guessing rate and lapsing rate are fixed, hence link is
a function of only one variable.
}
