% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_link.R
\name{probit_link}
\alias{probit_link}
\title{Probit link with guessing and lapsing rates}
\usage{
probit_link(guessing = 0, lapsing = 0)
}
\arguments{
\item{guessing}{(optional) guessing rate; default is 0}

\item{lapsing}{(optional) lapsing rate; default is 0}
}
\value{
link  probit link for use in all GLM functions
}
\description{
Creates a complementary log-log link function; the guessing rate and lapsing rate are fixed, hence link is
a function of only one variable.
}
